basic post-routes skeleton created

+++ routes/post-routes.js
@@ -6,37 +6,54 @@ const {} = require
 // /login
 
 router.post('/login', (req, res) => {
-
+  const email = req.body.email;
+  const password = req.body.password;
+  // check passwoprd
+  let userObject = getUserByEmail()
+    if (password === userObject.password) {
+    res.redirect('/passwords')
+  }
+  // add cookie
+  // TODO - write helper function cookie
+  // redirect to password dashboard
+  res.redirect('/passwords')
 });
 
 // /register
 
 router.post('/register', (req, res) => {
-
+  const email = req.params.email;
+  const password = req.params.password;;
+  addUser(email, password);
 });
 
 // /:id/delete
 
 router.post('/:id/delete', (req, res) => {
-  //const userId = req.id
-  // deleteUser(userID)
+  const userId = req.id;
+  deleteUser(userID);
 });
 
 // /passwords
 
 router.post('/passwords', (req, res) => {
-  // const userId = req.session.user_id
-  // const label = req.body.label
-  // const username = req.body.username
-  // const category = req.body.category
-  // const org_id = getUserById(userId)
-  // addPassword(userId, label, username, password, category, org_id)
+  const userId = req.session.user_id;
+  const label = req.body.label;
+  const username = req.body.usernamew;
+  const category = req.body.category;
+  const org_id = getUserById(userId);
+  addPassword(userId, label, username, password, category, org_id);
 });
 
-// /passwords/:id
+// Edit individual password
 
 router.post('/passwords/:id', (req, res) => {
-
+  const userId = req.session.user_id;
+  const label = req.body.label;
+  const username = req.body.usernamew;
+  const category = req.body.category;
+  const org_id = getUserById(userId);
+  editPassword(userId, label, username, password, category, org_id);
 });
 
 // /passwords/:id/delete
@@ -54,17 +71,23 @@ router.post('/orgs', (req, res) => {
   // TODO: how do we add users to org?
 });
 
-// /org/:id
+// Add Org
 
 router.post('/orgs/:id', (req, res) => {
-
+  const orgName = req.params.org_name;
+  addOrg(orgName)
 });
 
-// /orgs/:id/delete
+// Delete Org
 
 router.post('/orgs/:id/delete', (req, res) => {
-  const orgName = req.params.id
-  deleteOrg(orgName)
+  const userId = req.session.user_id;
+  const orgId = req.params.id
+  if (userIsOrgAdmin(userId, orgID)) {
+    deleteOrg(orgName);
+    return;
+  }
+  // Error - insufficient priledges
 });
 
 // /orgs/:id/userid
@@ -74,7 +97,8 @@ router.post('/orgs/:id/:userid', (req, res) => {
   const userId = req.cookies.user_id
   const isAdmin = false;
   if (isInOrg(userId, orgId)) {
-    updateUserInOrg(orgId, userID, isAdmin)
+    updateUserInOrg(orgId, userID, isAdmin);
+    return;
   }
   // TODO: how do we authenticate admin privs?
   addUserToOrg(orgId, userID, isAdmin)
